+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Ensure whitehall content tagged to mainstream browse and the topic taxonomy has a topic taxonomy breadcrumb ([PR #3871](https://github.com/alphagov/govuk_publishing_components/pull/3871))
 * Add GA4 tracking to the document list component ([PR #3874](https://github.com/alphagov/govuk_publishing_components/pull/3874))
 * Remove GA4 'action' from navigation content history events ([PR #3877](https://github.com/alphagov/govuk_publishing_components/pull/3877))
+* Fix empty input validation issues ([PR #3514](https://github.com/alphagov/govuk_publishing_components/pull/3514))
 
 ## 37.4.0